+++ src/scweaver/Starcounter.Internal.Weaver/DatabaseTypePolicy.cs
@@ -10,6 +10,7 @@ namespace Starcounter.Internal.Weaver {
     internal class DatabaseTypePolicy {
         DatabaseTypeConfiguration config;
         IType databaseAttributeType;
+        IType transientAttributeType;
 
         /// <summary>
         /// Gets the underlying <see cref="DatabaseTypeConfiguration"/> influencing
@@ -21,18 +22,31 @@ namespace Starcounter.Internal.Weaver {
             }
         }
 
-        public DatabaseTypePolicy(DatabaseTypeConfiguration typeConfiguration, IType databaseAttribute) {
+        public DatabaseTypePolicy(DatabaseTypeConfiguration typeConfiguration, IType databaseAttribute, IType transientAttribute) {
             config = typeConfiguration;
             databaseAttributeType = databaseAttribute;
+            transientAttributeType = transientAttribute;
         }
 
         public bool IsDatabaseType(TypeDefDeclaration typeDef) {
-            if (config.IsConfiguredDatabaseType(typeDef.Name)) {
-                if (typeDef.IsPublic()) {
+            // Explicit transient attribute overrides everything else
+            var transient = IsTaggedWithTransientAttribute(typeDef);
+            if (transient) return false;
+
+            // For public types, we allow them to be database classes by
+            // means of configuration / assembly level constructs. Check
+            // that first.
+            if (typeDef.IsPublic()) {
+                if (IsInDatabaseAttributeAssembly(typeDef)) {
+                    return true;
+                }
+
+                if (config.IsConfiguredDatabaseType(typeDef.Name)) {
                     return true;
                 }
             }
-
+            
+            // Finally the standard path
             return IsTaggedWithDatabaseAttribute(typeDef);
         }
 
@@ -56,9 +70,25 @@ namespace Starcounter.Internal.Weaver {
             });
         }
 
+        bool IsInDatabaseAttributeAssembly(TypeDefDeclaration typeDef) {
+            return typeDef.Module.AssemblyManifest.CustomAttributes.Contains(databaseAttributeType);
+        }
+
         bool IsTaggedWithDatabaseAttribute(TypeDefDeclaration typeDef) {
             var cursor = typeDef;
             while (!cursor.CustomAttributes.Contains(databaseAttributeType)) {
+                if (cursor.BaseType != null) {
+                    cursor = cursor.BaseType.GetTypeDefinition();
+                } else {
+                    return false;
+                }
+            }
+            return true;
+        }
+
+        bool IsTaggedWithTransientAttribute(TypeDefDeclaration typeDef) {
+            var cursor = typeDef;
+            while (!cursor.CustomAttributes.Contains(transientAttributeType)) {
                 if (cursor.BaseType != null) {
                     cursor = cursor.BaseType.GetTypeDefinition();
                 } else {